1.2.0 优化代码生成模板 支持动态任务标题配置 更新模型和说明

+++ bpmc/src/main/java/com/dili/bpmc/controller/HistoryController.java
@@ -184,7 +184,7 @@ public class HistoryController {
         EasyuiPageOutput easyuiPageOutput = new EasyuiPageOutput();
         List results = ValueProviderUtils.buildDataByProvider(param, list);
         easyuiPageOutput.setRows(results);
-        easyuiPageOutput.setTotal((int)total);
+        easyuiPageOutput.setTotal(total);
         return JSON.toJSONString(easyuiPageOutput, SerializerFeature.IgnoreErrorGetter);
     }
